% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{all_measures}
\alias{all_measures}
\title{Get all measures}
\usage{
all_measures(s, a = NULL, drop.r = NULL, maxOrder = 7)
}
\arguments{
\item{s}{Sequence vector.}

\item{a}{Optional. Vector of alternatives (e.g. 1:6 in a mental dice task).}

\item{drop.r}{Optional. Whether to drop repetitions when calculating other measures}
}
\value{
A dataframe with all measures in the package
}
\description{
Get all measures
}
\examples{
all_measures(sample(1:5, 100, T))
}
