% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{block_entropy}
\alias{block_entropy}
\title{Calculate block entropy (as in randfindR package)}
\usage{
block_entropy(x, block_size = 1L)
}
\arguments{
\item{x}{sequence}

\item{block_size}{length of blocks in which the original sequence should be divided for analysis}
}
\description{
Calculate block entropy (as in randfindR package)
}
