% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{lz76_complexity}
\alias{lz76_complexity}
\title{Calculate Lempel-Ziv complexity as implemented in Angelike & Musch, 2024}
\usage{
lz76_complexity(v)
}
\arguments{
\item{v}{sequence}
}
\description{
Calculate Lempel-Ziv complexity as implemented in Angelike & Musch, 2024
}
