% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{phi_index}
\alias{phi_index}
\title{Phi Index}
\usage{
phi_index(s, a = NULL, maxOrder = 7)
}
\arguments{
\item{s}{Sequence vector.}

\item{a}{Optional. Vector of alternatives (e.g. 1:6 in a mental dice task).}

\item{maxOrder}{Optional. Highest phi index to calculate}
}
\value{
vector of phi indices from 2 to maxOrder.
}
\description{
Phi Index
}
\examples{
phi_index(sample(1:20, 100, T))
}
