% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/randMeasures-package.R
\docType{package}
\name{randMeasures-package}
\alias{randMeasures}
\alias{randMeasures-package}
\title{randMeasures: Tools for analysis of human random sequence generation.}
\description{
This package provides a series of measures employed in the random generation literature to evaluate people's sequences.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://github.com/lucas-castillo/randMeasures}
  \item \url{https://lucas-castillo.github.io/randMeasures/}
  \item Report bugs at \url{https://github.com/lucas-castillo/randMeasures/issues}
}

}
\author{
\strong{Maintainer}: Lucas Castillo \email{lucas.castillo.marti@gmail.com} (\href{https://orcid.org/0000-0003-0274-0777}{ORCID})

}
\keyword{internal}
