% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{redundancy}
\alias{redundancy}
\title{Redundancy}
\usage{
redundancy(s, a = NULL)
}
\arguments{
\item{s}{Sequence vector.}

\item{a}{Optional. Vector of alternatives (e.g. 1:6 in a mental dice task).}
}
\value{
Redundancy value from 0 to 1.
}
\description{
Redundancy
}
\examples{
redundancy(sample(1:10, 100, T))
}
