% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/functions.R
\name{repetition_gap}
\alias{repetition_gap}
\title{Repetition Gap}
\usage{
repetition_gap(s, a = NULL, measure = "median")
}
\arguments{
\item{s}{Sequence vector.}

\item{a}{Optional. Vector of alternatives (e.g. 1:6 in a mental dice task).}

\item{measure}{String. Either 'median', 'mode' or 'mean'.}
}
\value{
Gap measure (either mode, median or mean, as specified by \code{measure}).
}
\description{
Repetition Gap
}
\examples{
v <- sample(1:100, 100, T)
repetition_gap(v, measure = "mean")
repetition_gap(v, measure = "median")
repetition_gap(v, measure = "mode")
}
