% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{RQA}
\alias{RQA}
\title{Recurrent Quantification Analyses}
\usage{
RQA(sequence, lmin = 2L, vmin = 2L)
}
\arguments{
\item{sequence}{sequence to analyse}

\item{lmin}{minimum diagonal length for DET}

\item{vmin}{minimum vertical length for LAM}
}
\value{
list of RQA results
}
\description{
Recurrent Quantification Analyses
}
