% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{SWLZ_ER}
\alias{SWLZ_ER}
\title{Calculate Lempel-Ziv entropy rate using sliding window method (Vegetabile et al., 2019)}
\usage{
SWLZ_ER(sequence)
}
\description{
Calculate Lempel-Ziv entropy rate using sliding window method (Vegetabile et al., 2019)
}
